Add tests
Move tensor tests in separate file
